requirements.txt
azure-identity==1.24.0
azure-ai-evaluation==1.10.0
azure-ai-projects==1.0.0
azure-core==1.34.0
requests==2.32.3
python-dotenv==1.1.0
openai==2.8.0
numpy==2.3.1